NN-5419: split adjudications API client into system and user token instances
